+++ src/main/java/ga/GeneticAlgorithm.java
@@ -6,8 +6,8 @@ public class GeneticAlgorithm {
 
     // Parameters
     private final int populationSize = 100;
-    private final double crossOverRate = 0.5;
-    private final double mutationRate = 0.01;
+    private final float crossOverRate = 0.5F;
+    private final float mutationRate = 0.01F;
     private final int durationPenaltyRate = 20;
     private final boolean elitism = true;
